[REFACTOR(HomeFragment)] 어댑터 프레젠터로 이동 및 바인딩적용
