+++ app/src/main/java/woowacourse/movie/ui/home/presenter/HomeContract.kt
@@ -1,15 +1,15 @@
 package woowacourse.movie.ui.home.presenter
 
-import woowacourse.movie.model.MovieListModel
 import woowacourse.movie.ui.common.BaseView
+import woowacourse.movie.ui.home.adapter.HomeAdapter
 
 interface HomeContract {
     interface View : BaseView<Presenter> {
-        var movieWithAdvertisement: List<MovieListModel>
+        var homeAdapter: HomeAdapter
     }
 
     interface Presenter {
 
-        fun getMovieWithAdvertisement()
+        fun initAdapter()
     }
 }
